// src/core/card/format.enum.ts
export enum Format {
    Standard = "standard",
    Commander = "commander",
    Modern = "modern",
    Legacy = "legacy",
    Vintage = "vintage",
    Brawl = "brawl",
    Explorer = "explorer",
    Historic = "historic",
    Oathbreaker = "oathbreaker",
    Pauper = "pauper",
    Pioneer = "pioneer",
}